add state SpecialStyleAppended
